WoT restructure: move stickyScrollStrategy scroll/ -> overlay/strategies/ (overlay-owned behavior)
